requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
langchain==0.1.9
langchain-openai==0.0.7
langchain-core>=0.1.26
httpx==0.25.1
python-dotenv==1.0.0
